requirements.txt
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
wsproto==1.1.0
pandas==1.2.5
numpy==1.20.1